Completed the res api module

+++ rest-app/backend/controllers/feed.js
@@ -123,9 +123,15 @@ exports.updatePost = (req, res, next) => {
 				error.statusCode = 400;
 				throw error;
 			}
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error("Unauthorized");
+				error.statusCode = 403;
+				throw error;
+			}
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl);
 			}
+
 			post.title = title;
 			post.content = content;
 			post.imageUrl = imageUrl;
@@ -152,8 +158,20 @@ exports.deletePost = (req, res, next) => {
 				throw error;
 			}
 			// Check logged in user
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error("Unauthorized");
+				error.statusCode = 403;
+				throw error;
+			}
 			clearImage(post.imageUrl);
 			return Post.findByIdAndRemove(postId);
+		})
+		.then((result) => {
+			return User.findById(req.userId);
+		})
+		.then((user) => {
+			user.posts.pull(postId);
+			return user.save();
 		})
 		.then((result) => {
 			console.log(result);
